+++ examples/api_integrations/ben_bot.py
@@ -30,6 +30,7 @@ class MyClient(fortnitepy.Client):
                 email=email,
                 password=password,
                 prompt_authorization_code=True,
+                prompt_code_if_invalid=True,
                 delete_existing_device_auths=True,
                 **device_auth_details
             )