+++ fortnitepy/stats.py
@@ -117,7 +117,7 @@ class StatsV2(_StatsBase):
 
             # gets ninjas kd in solo on input touch
             async def get_ninja_touch_solo_kd():
-                user = await client.fetch_user('Ninja)
+                user = await client.fetch_user('Ninja')
                 stats = await client.fetch_br_stats(user.id)
 
                 return stats.get_kd(stats.get_stats()['touch']['defaultsolo'])